LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-top-right-controls-bar-hidden-when-mute-button-disabled.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the top right controls bar isn't visible when the mute button is disabled.");

const mediaControls = new MacOSInlineMediaControls({ width: 300, height: 250 });
document.body.appendChild(mediaControls.element);

mediaControls.element.style.position = 'absolute';
mediaControls.element.style.left = '-9999px';

mediaControls.layout();

(async function () {
    debug("Mute button is enabled");

    debug("mediaControls.topRightControlsBar.parent.constructor.name: " + mediaControls.topRightControlsBar.parent.constructor.name);

    debug("")
    debug("Mute button is disabled");
    mediaControls.muteButton.enabled = false;
    mediaControls.layout();
    
    debug("mediaControls.muteButton.parent.parent.constructor.name: " + (mediaControls.muteButton.parent ? mediaControls.muteButton.parent.parent.constructor.name : "null"));
    shouldBeNull("mediaControls.topRightControlsBar.parent");
    
    mediaControls.element.remove();

    debug("");

    finishJSTest();
})();

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
