LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-top-right-controls-bar-hidden-when-mute-button-disabled-expected.txt

Testing the top right controls bar isn't visible when the mute button is disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Mute button is enabled
mediaControls.topRightControlsBar.parent.constructor.name: MacOSInlineMediaControls

Mute button is disabled
mediaControls.muteButton.parent.parent.constructor.name: MacOSInlineMediaControls
PASS mediaControls.topRightControlsBar.parent is null

PASS successfullyParsed is true

TEST COMPLETE
